fs, pstore: fix warning
warning: comparison of distinct pointer types lacks a cast.

Reviewed-on: https://opticm6.rds.intel.com/r/73131
Tested-by: TCloud Jenkins

+++ fs/pstore/platform.c
@@ -377,7 +377,8 @@ static void pstore_dump(struct kmsg_dumper *dumper,
 			while (pstore_extra_written < pstore_extra_size) {
 				size_t read;
 				size_t len = min(psinfo->bufsize,
-						pstore_extra_size - pstore_extra_written);
+						(size_t)pstore_extra_size-
+						pstore_extra_written);
 
 				ret = extra_dumper->get_data(psinfo->buf, len, &read,
 						extra_dumper->priv);
